requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
alembic==1.18.4
celery==5.6.3
redis==7.4.0
stripe==15.0.1
httpx==0.28.1
reportlab==4.4.1
weasyprint==68.1
python-dotenv==1.1.0
pydantic-settings==2.13.1
jinja2==3.1.6
sendgrid==6.12.5
boto3==1.42.89
geopy==2.4.1
email-validator==2.2.0
shapely==2.1.2
pyproj==3.7.2
